+++ email-notification-service/src/middleware.ts
@@ -38,7 +38,7 @@ export const verifyUserJWT = (req: Request, res, next): void => {
     });
 };
 export const verifyNotificationJWT = (req: Request, res, next): void => {
-  logger.trace("Verifying Notification-JWT ...");
+  logger.debug("Verifying Notification-JWT ...");
   const token: string = getJWTToken(req);
 
   verifyJWTToken(token)
@@ -47,8 +47,14 @@ export const verifyNotificationJWT = (req: Request, res, next): void => {
       next();
     })
     .catch((err) => {
-      logger.trace({ err, token }, "Notification-JWT invalid");
-      const body: InvalidJWTResponseBody = { message: "Invalid JWT token provided." };
+      logger.error({ err, token }, "Notification-JWT invalid");
+      let message = "Invalid JWT token provided.";
+      if (err.message === "invalid algorithm") {
+        message = err.message;
+      } else if (err.message === "jwt expired") {
+        message = err.message;
+      }
+      const body: InvalidJWTResponseBody = { message };
       res.status(400).json(body);
     });
 };