+++ frontend/src/languages/german.js
@@ -18,7 +18,7 @@ const de = {
     add_without_budget: "Hinzufügen ohne Budget",
     added: "Hinzugefügt",
     additional_data: "Zusätzliche Daten",
-    amount: "Betrag",
+    amount: "Finanzierungsbetrag",
     approver: "Genehmiger",
     assign: "Zuweisen",
     assigned: "Zugewiesen",
@@ -225,7 +225,7 @@ const de = {
       "Nur Workflow-Elemente vom Typ 'eingeschränkt' zulassen. Bei Zuweisung eines eingeschränkten Workflow-Items an einen anderen User werden Berechtigungen automatisch erteilt und entzogen. Der Zuweisende behält nur die Anzeigerechte.",
     workflowitem_assignee: "Fester Zuständiger",
     organization_info: "Förderer",
-    total_budget_info: "Total budget",
+    total_budget_info: "Beteiligter Stakeholder",
     default_assignee_warning: "Fixed assignee cannot be changed once Subproject is created.",
     default_assignee_warning2:
       "Fixed assignee will be assigned to all workflow items in subproject without an option to change it."
@@ -561,6 +561,7 @@ const de = {
     workflowitem_intent_revokePermission: "Die Berechtigungen für Workflow action {0} wurden geändert",
     workflowitem_update: "Workflow action {0} wurde aktualisiert",
     workflowitem_updated: "Workflow action {0} wurde aktualisiert",
+    workflowitem_created: "Workflow action {0} wurde Ihnen zugewiesen",
     payload_error_message:
       "Ups! ... Es ist nicht deine Schuld - Ein clientseitiger Validierungsfehler ist aufgetreten. Bitte informieren Sie den Administrator."
   },
@@ -653,6 +654,7 @@ const de = {
     workflowitem_intent_grantPermission: "Workflow action-Erlaubnis erteilen",
     workflowitem_intent_listPermissions: "Workflow action-Berechtigungen anzeigen",
     workflowitem_intent_revokePermission: "Workflow action-Erlaubnis widerrufen",
+    workflowitem_list: "Workflow-Eintrag anzeigen",
     workflowitem_update: "Workflow-Eintrag aktualisieren",
     workflowitem_view: "Workflow-Eintrag anzeigen",
     write: "Schreibberechtigungen"