+++ frontend/src/languages/portuguese.js
@@ -19,7 +19,7 @@ const pt = {
     add_without_budget: "Adicionar sem orçamento",
     added: "Adicionado",
     additional_data: "Dados Adicionais",
-    amount: "Valor",
+    amount: "Valor do Financiamento",
     approver: "Aprovador",
     assign: "Atribuir",
     assigned_budget: "Orçamento atribuído",
@@ -312,7 +312,7 @@ const pt = {
       "Permitir apenas item de fluxo de trabalho do tipo restrito. Ao atribuir um item de fluxo de trabalho restrito, as permissões são concedidas e revogadas automaticamente. O atribuidor manterá apenas as permissões de visualização.",
     workflowitem_assignee: "Cessionário padrão",
     organization_info: "Organização financiadora",
-    total_budget_info: "Total budget",
+    total_budget_info: "Parte interessada participante",
     default_assignee_warning: "Fixed assignee cannot be changed once Subproject is created.",
     default_assignee_warning2:
       "Fixed assignee will be assigned to all workflow actions items in subproject without an option to change it."
@@ -561,6 +561,7 @@ const pt = {
     workflowitem_intent_revokePermission: "As permissões para o item de workflow action {0} foram alteradas",
     workflowitem_update: "Item de Workflow action {0} foi atualizado",
     workflowitem_updated: "Item de Workflow action {0} foi atualizado",
+    workflowitem_created: "Item de workflow action {0} foi atribuído a você",
     payload_error_message:
       "Ops! ... Não é sua culpa - Ocorreu um erro de validação do lado do cliente. Informe o administrador."
   },
@@ -653,8 +654,9 @@ const pt = {
     workflowitem_intent_grantPermission: "Conceder permissões do item de workflow action",
     workflowitem_intent_listPermissions: "Visualizar permissões do item de workflow action",
     workflowitem_intent_revokePermission: "Revogar permissões do item de workflow action",
-    workflowitem_update: "Atualizar item de worflow",
-    workflowitem_view: "Visualizar item de workflow action",
+    workflowitem_list: "Visualizar action de workflow",
+    workflowitem_update: "Atualizar action de worflow",
+    workflowitem_view: "Visualizar action de workflow action",
     write: "Permissões de escrita"
   },