+++ frontend/src/pages/Workflows/WorkflowDialog.js
@@ -31,7 +31,8 @@ const handleCreate = (props) => {
     exchangeRate,
     dueDate,
     workflowitemType,
-    tags
+    tags,
+    fundingOrganization
   } = workflowToAdd;
 
   createWorkflowItem(
@@ -47,7 +48,8 @@ const handleCreate = (props) => {
     workflowitemType,
     projectDisplayName,
     subprojectDisplayName,
-    tags
+    tags,
+    fundingOrganization
   );
 
   storeSnackbarMessage(strings.formatString(strings.snackbar.permissions_warning, shortenedDisplayName(displayName)));
@@ -75,7 +77,8 @@ const handleCreateFromTemplate = (props) => {
     dueDate,
     exchangeRate,
     status,
-    workflowitemType
+    workflowitemType,
+    fundingOrganization
   } = workflowToAdd;
 
   const workflowitems = createWorkflowItemsFromTemplate(workflowTemplate, workflowToAdd);
@@ -94,7 +97,8 @@ const handleCreateFromTemplate = (props) => {
     subprojectDisplayName,
     workflowitemType,
     workflowTemplate,
-    workflowitems
+    workflowitems,
+    fundingOrganization
   });
 
   storeSnackbarMessage(strings.formatString(strings.snackbar.permissions_warning, shortenedDisplayName(displayName)));
@@ -110,10 +114,12 @@ const handleEdit = (props) => {
       delete workflowToAdd.amount;
       delete workflowToAdd.currency;
       delete workflowToAdd.exchangeRate;
+      delete workflowToAdd.fundingOrganization;
     } else {
       workflowToAdd.amount = "";
       workflowToAdd.currency = "";
       workflowToAdd.exchangeRate = 1;
+      workflowToAdd.fundingOrganization = "";
     }
   }
   const changes = compareWorkflowItems(originalWorkflowItem, workflowToAdd);
@@ -140,8 +146,18 @@ const handleEdit = (props) => {
 };
 
 const createWorkflowItemsFromTemplate = (selectedTemplate, workflowToAdd) => {
-  const { amount, amountType, currency, description, documents, dueDate, exchangeRate, status, workflowitemType } =
-    workflowToAdd;
+  const {
+    amount,
+    amountType,
+    currency,
+    description,
+    documents,
+    dueDate,
+    exchangeRate,
+    status,
+    workflowitemType,
+    fundingOrganization
+  } = workflowToAdd;
   const template = templates[selectedTemplate];
   return template.steps.map((step) => {
     return {
@@ -154,7 +170,8 @@ const createWorkflowItemsFromTemplate = (selectedTemplate, workflowToAdd) => {
       dueDate,
       exchangeRate,
       status,
-      workflowitemType
+      workflowitemType,
+      fundingOrganization
     };
   });
 };
@@ -181,7 +198,10 @@ const WorkflowDialog = (props) => {
     setStorageServiceAvailable,
     workflowTemplate,
     dialogTitle,
-    hideWorkflowDialog
+    hideWorkflowDialog,
+    deleteDocument,
+    deleteWorkflowDocument,
+    deleteWorkflowDocumentExternalLink
   } = props;
 
   useEffect(() => {
@@ -234,10 +254,13 @@ const WorkflowDialog = (props) => {
         handleSubmit: handleCreate,
         dialogShown: creationDialogShown
       };
-  const { displayName, amountType, amount } = workflowToAdd;
+  const { displayName, amountType, amount, fundingOrganization, id } = workflowToAdd;
+  const { location } = props;
   const exchangeRate = fromAmountString(workflowToAdd.exchangeRate);
   const changes = compareObjects(workflowItems, workflowToAdd);
   delete changes.assignee;
+  const projectId = location.pathname.split("/")[2];
+  const subprojectId = location.pathname.split("/")[3];
   const documentStep = {
     title: strings.workflow.workflow_documents,
     content: (
@@ -245,12 +268,21 @@ const WorkflowDialog = (props) => {
         storeWorkflowDocument={storeWorkflowDocument}
         storeWorkflowDocumentExternalLink={storeWorkflowDocumentExternalLink}
         workflowDocuments={workflowToAdd.documents}
+        deleteDocument={deleteDocument}
+        deleteWorkflowDocument={deleteWorkflowDocument}
+        deleteWorkflowDocumentExternalLink={deleteWorkflowDocumentExternalLink}
+        projectId={projectId}
+        subprojectId={subprojectId}
+        workflowitemId={id}
         {...props}
       />
     ),
     nextDisabled:
-      workflowToAdd.amountType === "N/A" && Object.keys(changes).length === 2
-        ? Object.keys(changes).length === 2 && Object.hasOwn(changes, "currency") && Object.hasOwn(changes, "amount")
+      workflowToAdd.amountType === "N/A" && Object.keys(changes).length === 3
+        ? Object.keys(changes).length === 3 &&
+          Object.hasOwn(changes, "currency") &&
+          Object.hasOwn(changes, "amount") &&
+          Object.hasOwn(changes, "fundingOrganization")
         : _isEmpty(changes)
   };
 
@@ -259,7 +291,7 @@ const WorkflowDialog = (props) => {
       title: strings.workflow.workflow_name,
       nextDisabled:
         (workflowTemplate === "" && _isEmpty(displayName)) ||
-        (amountType !== "N/A" && amount === "") ||
+        (amountType !== "N/A" && (amount === "" || fundingOrganization === "")) ||
         (amountType !== "N/A" && (!Number.isFinite(exchangeRate) || exchangeRate === 0)),
       content: (
         <div className="workflow-dialog-content">