+++ frontend/src/pages/Workflows/actions.js
@@ -35,7 +35,10 @@ export const WORKFLOW_DUEDATE = "WORKFLOW_DUEDATE";
 export const WORKFLOW_ASSIGNEE = "WORKFLOW_ASSIGNEE";
 export const WORKFLOW_DOCUMENT = "WORKFLOW_DOCUMENT";
 export const WORKFLOW_DOCUMENT_EXTERNAL_LINK = "WORKFLOW_DOCUMENT_EXTERNAL_LINK";
+export const DELETE_WORKFLOW_DOCUMENT_EXTERNAL_LINK = "DELETE_WORKFLOW_DOCUMENT_EXTERNAL_LINK";
+export const DELETE_WORKFLOW_DOCUMENT = "DELETE_WORKFLOW_DOCUMENT";
 export const WORKFLOWITEM_TYPE = "WORKFLOWITEM_TYPE";
+export const WORKFLOW_FUNDING_ORGANIZATION = "WORKFLOW_FUNDING_ORGANIZATION";
 export const CREATE_WORKFLOW = "CREATE_WORKFLOW";
 export const CREATE_WORKFLOW_FROM_TEMPLATE = "CREATE_WORKFLOW_FROM_TEMPLATE";
 export const CREATE_WORKFLOW_FAILURE = "CREATE_WORKFLOW_FAILURE";
@@ -495,7 +498,8 @@ export function showEditDialog(
   documents,
   dueDate,
   workflowitemType,
-  tags
+  tags,
+  fundingOrganization
 ) {
   return {
     type: SHOW_WORKFLOW_EDIT,
@@ -509,7 +513,8 @@ export function showEditDialog(
     documents,
     dueDate,
     workflowitemType,
-    tags
+    tags,
+    fundingOrganization
   };
 }
 
@@ -550,6 +555,20 @@ export function storeWorkflowDocumentExternalLink(link, fileName, linkedFileHash
     fileName,
     linkedFileHash,
     comment
+  };
+}
+
+export function deleteWorkflowDocumentExternalLink(linkedFileHash) {
+  return {
+    type: DELETE_WORKFLOW_DOCUMENT_EXTERNAL_LINK,
+    linkedFileHash
+  };
+}
+
+export function deleteWorkflowDocument(base64) {
+  return {
+    type: DELETE_WORKFLOW_DOCUMENT,
+    base64
   };
 }
 
@@ -597,6 +616,13 @@ export function storeWorkflowCurrency(currency) {
   return {
     type: WORKFLOW_CURRENCY,
     currency: currency
+  };
+}
+
+export function storeWorkflowFundingOrga(fundingOrganization) {
+  return {
+    type: WORKFLOW_FUNDING_ORGANIZATION,
+    fundingOrganization
   };
 }
 
@@ -652,7 +678,8 @@ export function createWorkflowItemAction(
   subprojectDisplayName,
   assignee,
   assigneeDisplayName,
-  tags
+  tags,
+  fundingOrganization
 ) {
   return {
     type: CREATE_WORKFLOW,
@@ -672,7 +699,8 @@ export function createWorkflowItemAction(
     subprojectDisplayName,
     assignee,
     assigneeDisplayName,
-    tags
+    tags,
+    fundingOrganization
   };
 }